+++ packages/flutter_driver/lib/flutter_driver.dart
@@ -13,11 +13,10 @@
 /// Protractor (Angular), Espresso (Android) or Earl Gray (iOS).
 library flutter_driver;
 
-export 'src/common/command_deserializer_factory.dart';
-export 'src/common/create_finder_factory.dart';
 export 'src/common/diagnostics_tree.dart';
 export 'src/common/enum_util.dart';
 export 'src/common/error.dart';
+export 'src/common/factory.dart';
 export 'src/common/find.dart';
 export 'src/common/frame_sync.dart';
 export 'src/common/fuchsia_compat.dart';
