--- packages/flutter_driver/lib/src/common/command_deserializer_factory.dart
@@ -1,7 +1,0 @@
-// Copyright 2014 The Flutter Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-class CommandDeserializerFactory {
-
-}
